+++ R/as_fc.R
@@ -13,6 +13,8 @@
 #' @param text_padding Changes the text padding inside the box. Default is 1. This number has to be greater than 0.
 #' @param bg_fill Box background color. It is white by default. See the `fill` parameter for \code{\link{gpar}}.
 #' @param border_color Box border color. It is black by default. See the `col` parameter for \code{\link{gpar}}.
+#' @param width Width of the box. If NA, it automatically adjusts to the content (default). Must be an object of class \code{\link{unit}} or a number between 0 and 1.
+#' @param height Height of the box. If NA, it automatically adjusts to the content (default). Must be an object of class \code{\link{unit}} or a number between 0 and 1.
 #' @param hide Logical value to hide the initial box or not. Default is FALSE. hide = TRUE can only be combined with fc_split().
 #'
 #' @return List with the dataset and the initialized flowchart parameters.
@@ -24,7 +26,7 @@
 #'
 #' @export
 
-as_fc <- function(.data = NULL, N = NULL, label = "Initial dataframe", text_pattern = "{label}\n{N}", just = "center", text_color = "black", text_fs = 8, text_fface = 1, text_ffamily = NA, text_padding = 1, bg_fill = "white", border_color = "black", hide = FALSE) {
+as_fc <- function(.data = NULL, N = NULL, label = "Initial dataframe", text_pattern = "{label}\n{N}", just = "center", text_color = "black", text_fs = 8, text_fface = 1, text_ffamily = NA, text_padding = 1, bg_fill = "white", border_color = "black", width = NA, height = NA, hide = FALSE) {
 
   if(is.null(.data) & is.null(N)) {
     cli::cli_abort("Either {.arg .data} or {.arg N} arguments must be specified.")
@@ -40,10 +42,6 @@ as_fc <- function(.data = NULL, N = NULL, label = "Initial dataframe", text_patt
 
   if(text_padding == 0) {
     cli::cli_abort("Text padding cannot be equal to zero.")
-  }
-
-  if(text_pattern != "{label}\n{N}") {
-
   }
 
   if(!hide) {
@@ -64,7 +62,9 @@ as_fc <- function(.data = NULL, N = NULL, label = "Initial dataframe", text_patt
       text_ffamily = text_ffamily,
       text_padding = text_padding,
       bg_fill = bg_fill,
-      border_color = border_color
+      border_color = border_color,
+      width = width,
+      height = height
     )
 
     if(is.character(label)) {